// src/services/api/generated/model/recipeDetailResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Stock API
 * API de gestion de stock (produits, catégories, recettes, utilisateurs)
 * OpenAPI spec version: v0.0.1
 */
import type { RecipeIngredientResponse } from './recipeIngredientResponse';

export interface RecipeDetailResponse {
  id?: string;
  name?: string;
  ingredients?: RecipeIngredientResponse[];
}