// src/services/api/generated/model/recipeSummaryResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Stock API
 * API de gestion de stock (produits, catégories, recettes, utilisateurs)
 * OpenAPI spec version: v0.0.1
 */

export interface RecipeSummaryResponse {
  id?: string;
  name?: string;
  ingredientCount?: number;
  createdAt?: string;
}