// src/services/api/generated/model/stockItemResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Stock API
 * API de gestion de stock (produits, catégories, recettes, utilisateurs)
 * OpenAPI spec version: v0.0.1
 */
import type { StockItemProductSummary } from './stockItemProductSummary';
import type { StockItemResponseStatus } from './stockItemResponseStatus';

export interface StockItemResponse {
  id?: string;
  product?: StockItemProductSummary;
  quantity?: number;
  lowThreshold?: number;
  expirationDate?: string;
  status?: StockItemResponseStatus;
  needsQuantityUpdate?: boolean;
}